Move MongoJsonObjectListField to the file containing JsonObjectField, which is consistent with how other such fields are organized
