+++ persistence/mongodb-record/src/main/scala/net/liftweb/mongodb/record/field/JsonObjectField.scala
@@ -16,13 +16,14 @@ package mongodb
 package record
 package field
 
-import scala.xml.{NodeSeq, Text}
+import scala.xml.NodeSeq
 import net.liftweb.common.{Box, Empty, Failure, Full}
-import net.liftweb.http.js.JE.{JsNull, Str}
 import net.liftweb.json._
-import net.liftweb.record.{Field, FieldHelpers, MandatoryTypedField, Record}
+import net.liftweb.record.{Field, FieldHelpers, MandatoryTypedField}
 import net.liftweb.util.Helpers.tryo
-import com.mongodb.DBObject
+import com.mongodb.{BasicDBList, DBObject}
+
+import scala.collection.JavaConverters._
 
 abstract class JsonObjectField[OwnerType <: BsonRecord[OwnerType], JObjectType <: JsonObject[JObjectType]]
   (rec: OwnerType, valueMeta: JsonObjectMeta[JObjectType])
@@ -82,3 +83,31 @@ abstract class JsonObjectField[OwnerType <: BsonRecord[OwnerType], JObjectType <
 
 }
 
+/*
+* List of JsonObject case classes
+*/
+class MongoJsonObjectListField[OwnerType <: BsonRecord[OwnerType], JObjectType <: JsonObject[JObjectType]]
+(rec: OwnerType, valueMeta: JsonObjectMeta[JObjectType])(implicit mf: Manifest[JObjectType])
+  extends MongoListField[OwnerType, JObjectType](rec: OwnerType) {
+
+  override def asDBObject: DBObject = {
+    val dbl = new BasicDBList
+    value.foreach { v => dbl.add(JObjectParser.parse(v.asJObject()(owner.meta.formats))(owner.meta.formats)) }
+    dbl
+  }
+
+  override def setFromDBObject(dbo: DBObject): Box[List[JObjectType]] =
+    setBox(Full(dbo.keySet.asScala.toList.map(k => {
+      valueMeta.create(JObjectParser.serialize(dbo.get(k.toString))(owner.meta.formats).asInstanceOf[JObject])(owner.meta.formats)
+    })))
+
+  override def asJValue: JValue = JArray(value.map(_.asJObject()(owner.meta.formats)))
+
+  override def setFromJValue(jvalue: JValue) = jvalue match {
+    case JNothing|JNull if optional_? => setBox(Empty)
+    case JArray(arr) => setBox(Full(arr.map( jv => {
+      valueMeta.create(jv.asInstanceOf[JObject])(owner.meta.formats)
+    })))
+    case other => setBox(FieldHelpers.expectedA("JArray", other))
+  }
+}
