+++ persistence/mongodb-record/src/main/scala/net/liftweb/mongodb/record/field/MongoListField.scala
@@ -19,20 +19,16 @@ package mongodb
 package record
 package field
 
+import com.mongodb._
+import net.liftweb.common.{Box, Empty, Failure, Full}
+import net.liftweb.http.SHtml
+import net.liftweb.json._
+import net.liftweb.record.{Field, FieldHelpers, MandatoryTypedField}
+import net.liftweb.util.Helpers._
+import org.bson.Document
+
 import scala.collection.JavaConversions._
 import scala.xml.NodeSeq
-
-import common.{Box, Empty, Failure, Full}
-import http.SHtml
-import http.js.JE.{JsNull, JsRaw}
-import json._
-import net.liftweb.record.{Field, FieldHelpers, MandatoryTypedField, Record}
-import util.Helpers._
-
-import com.mongodb._
-import org.bson.Document
-import org.bson.types.ObjectId
-import org.joda.time.DateTime
 
 /**
   * List field.
@@ -165,33 +161,4 @@ class MongoListField[OwnerType <: BsonRecord[OwnerType], ListType: Manifest](rec
     throw new RuntimeException("Warning, setting Document as field with no conversion, probably not something you want to do")
   }
 
-}
-
-/*
-* List of JsonObject case classes
-*/
-class MongoJsonObjectListField[OwnerType <: BsonRecord[OwnerType], JObjectType <: JsonObject[JObjectType]]
-  (rec: OwnerType, valueMeta: JsonObjectMeta[JObjectType])(implicit mf: Manifest[JObjectType])
-  extends MongoListField[OwnerType, JObjectType](rec: OwnerType) {
-
-  override def asDBObject: DBObject = {
-    val dbl = new BasicDBList
-    value.foreach { v => dbl.add(JObjectParser.parse(v.asJObject()(owner.meta.formats))(owner.meta.formats)) }
-    dbl
-  }
-
-  override def setFromDBObject(dbo: DBObject): Box[List[JObjectType]] =
-    setBox(Full(dbo.keySet.toList.map(k => {
-      valueMeta.create(JObjectParser.serialize(dbo.get(k.toString))(owner.meta.formats).asInstanceOf[JObject])(owner.meta.formats)
-    })))
-
-  override def asJValue: JValue = JArray(value.map(_.asJObject()(owner.meta.formats)))
-
-  override def setFromJValue(jvalue: JValue) = jvalue match {
-    case JNothing|JNull if optional_? => setBox(Empty)
-    case JArray(arr) => setBox(Full(arr.map( jv => {
-      valueMeta.create(jv.asInstanceOf[JObject])(owner.meta.formats)
-    })))
-    case other => setBox(FieldHelpers.expectedA("JArray", other))
-  }
 }
